+++ Backend/src/main/java/com/revature/data/GenericDAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface GenericDAO <T>{
     public T add(T t);
-    public T getById(Integer id);
+    public T getById(Long id);
     public List<T> getAll();
     public void update(T t);
     public void delete(T t);
